Revert "Add a sourceset field to LintModelModuleLibrary"
This reverts commit f6003a8510c2dab486bb7c292290c8c3f496198e.

Reason for revert: Moving from using a sourceset field to rely on the
artifact address to distinguish between different sourceset artifacts.
There is no need currently for lint to know whether a library or a
module is a testFixtures component of a library or a module. By
moving to using the artifact address lint will deal with them as
separate artifacts.

Change-Id: Iebd9569b46deebed75a17c7d60d75721bb3bfb04