[PSD] Quiet a run-time warning
The ActionToolbar platform implementation throws a warning if a toolbar
is created with a non-standard place.  This warning is logged every time
the toolbar is created, which in the Project Structure Dialog's case can
be every time the dependencies or modules tab is opened.  Quiet the
warning by explicitly setting the toolbar's target component to null.

Bug: n/a
Test: manual
Change-Id: I0cf85ddc56d27ad922cd5df0ba34039f60dca84d