[Sync] Throw on mismatched Canary versions
When we detect a mismatched AGP version, throw an
AndroidSyncException (and not an UnsupportedVersionException, which is
implicitly for Gradle versions, nor an AgpUpgradeRequiredException,
which requires properties that are not available in the model provider
worker).

Throwing an AndroidSyncException is I think the right thing, but leads
to appending the stack trace to the message because of issues with
serializing exceptions across different JVM versions.  This means that
we need to use find() not matches() to detect the error message
corresponding to this particular issue.

If we do have an incompatible version, instantiate and display the modal
forced-upgrade dialog.  Also generate a quickfix to re-instantiate the
forced-upgrade dialog in case the user cancels out of it before
realising that they cannot do very much in the resulting IDE.

Bug: 206397060
Bug: 206666447
Test: manual, added unit test
Change-Id: I80e97403c2df5e703c140bd099ed743c064fb722