Refactor expand all and collapse all actions
This is a preparation for implementing a tree table component tree
implementation. The Tree instance used in a TreeTable is not in the
Swing component hierarchy - so the tree expand and collapse all
would not work (the tree is not tecnically visible).

After this change these actions will work for both tree
implementations.

Bug: N/A
Test: N/A
Change-Id: Ie1cfb4a2112e413195d610a5605ebe5911c16606