Open files from DeviceExplorerFileManagerImpl
This change changes how Device Explorer opens files in the
editor. Essentially reverting the behaviour to how it was
before ag/9474023.

An external file opener is kept as optional so clients
of DeviceExplorer can override the default behaviour.
eg. Game-Tools.

The problem with the current implementation is that when
editor tabs of downloaded files are closed, the files are
deleted. But the logic that checks that those files come
from the editor is prone to false positives, which ends
up deleting files that should not be deleted.

The old implementation (which this CL brings back) keeps
track of files opened in the editor by Device Explorer,
which removes the risk of false positive when deleting
files.

Test: DeviceExplorerControllerTest
Bug: 202931544
Change-Id: I38b3bb75691f844b85d19070ca521991e84a3eb1