Break MigrateToAppCompatProcessor dependency on Gradle related code
`MigrateToAppCompatProcessor` is a build system agnostic refactoring
which used to depend on Gradle specific methods to get the current
configuration data.

This change replaces the APIs used to fetch the configuration data with
similar non-Gradle specific ones.

This change may impact the behavior in the case of a project in a non-
synced state, however it is doubtful that in this state the refactoring
can work well anyway.

Bug: 147586218
Test: existing MigrateToAppCompatGradleTest
Change-Id: I2c4fe60808c5a73c7618e6a685e7d15159fe8fe1