[Dsl] New transform for plugin models
In some forms of plugin application, for example

    apply plugin: 'name'

or

    apply {
      plugin 'name'
    }

there is no syntactically-legal way of expressing a version or apply
property.  Falling through to the default transform, however, is wrong,
because that default transform simply returns the Dsl element itself,
and then using the resulting model to set the value of the property
would lead to unexpected consequences.

Instead, define a transform which is always applicable and always
returns null (meaning "missing") for the transformed element, and which
throws on attempting to set the value.  This will at least prevent
client code from accidentally destructively modifying users' plugin
specifications.

Bug: none
Test: existing, more in subsequent CL
Change-Id: Iade94ded43d75d86a9204c45ecd7146391d62ee2