Fix AndroidRemoteDebugProcessHandler
Start notify AndroidProcessHandlerForDevices manually if we're going to attach debugger. It is needed for future work of processHandler.destroyProcess().

Usually ExecutionManagerImpl notifies ProcessHandler. But it's going to notify only AndroidRemoteDebugProcessHandler, because there should be 1 process handler for Execution.

We have two process handlers in debug case - one for app process on devices and another for debugger.

Fixes: 207119007
Test: manually
Change-Id: I628f2b815c6154beda4c929e62ac36e3eb491c11