[BA] set jetifier flag to false instead of removal
Also highlight the property line with the blinker (same way as it is
done for usages search).
Also add baloons with the feedback on this action. Show an info baloon
next to the modified property after navigating to the file with the
message that it was set to false and suggestion to consider removing it.
In case property is not found for some reason (e.g. it was already
removed) show error baloon with the corresponding message next to
action trigger (button or link).

Bug: 206821412
Test: existing tests
Change-Id: I2804fb804b5c04c3ec8e63efaf72cc9816b5d2ab