Re-enable KaptProjectResolverExtension
This is used to support kapt for non-androird modules. This change also
removes the KaptGradleModel from being visible for Android modules since
it messes up with our setup. This is due to the module names not being
the same as the source set names that we are obtaining from Gradle.

Fixes: 207389479
Test: snapshots updated
Change-Id: Icd25d4c9b5e4bd10020323da5c5f8dd8ea6aee18