[Dsl] Detect circularities in followElement()
It is possible to construct circularities in the dsl model's view of
things by using project property references; the parser/interpreter
doesn't currently have a good enough model of time to be able to deduce
which of the various possibilities a particular project reference might
refer to.  It is better to abandon trying to follow the elements rather
than guess or (worse) loop infinitely.

Bug: 197959708
Test: added
Change-Id: I36bfce2659cfa37ed5563103da37c95bdc85309d

GitOrigin-RevId: e31e5ebeff54c6f5a898db571275bc189b15611d