Add History Dropdown
Current text is added to history if:
* Enter is pressed
* The combo is opened

This is consistent with the way it works with original Logcat.

Unlike original Logcat, the dropdown is implemented as a pure ComboBox. In the original logcat we use a TextField with a UI plugin that adds a dropdown menu on the left.

We're using an EditorTextField wrapped in EditorComboBox which results in UI that looks more like the UI in the Debugger Evaluate dialog where an Editor is also used.

Bug: 191880898
Test: Added
Change-Id: I65dd5f4561b706f54da16d114feb08412ed310cb