Double force-stop on launch for APIs <= 25
On devices with APIs <= 25, there is a bug where the first shell
invocation of "am force-stop" will not terminate the target app,
and instead will place the app in the background. This causes
the app termination detector to not be able to not detect that
the app has been terminated, and thus timing out the launch.

The workaround is to double "am force-stop" when the the target
device is API <= 25, as the second force-stop will properly
terminate the application.

Fixes: 181004316
Test: LaunchTaskRunnerTest
Change-Id: I1b5a3cbb1bab3c9555f92347f16e324e3cc4a97a