Do not trigger upgrade assistant QF when unit testing.
This changes where we check for unit testing mode before triggering the
upgrade assistant QuickFixes.
we have an issueChecker that reports AGP versions that are below the
minimum supported by AS, and we recently added the possibility to
trigger upgrade assistant from it.

We do not want to trigger upgrade assistant in unit testing mode, but
we also want to be able to verify that we do get the expected behaviour
by AS when using an incompatible AGP version, and in such case, we
should be able to report to users that they are using an incompatible
AGP version, regardless of if we want to trigger the upgrade or not.

Bug: N/A
Test: following
Change-Id: Ic8ce5b762f7aff01e08f6f773ea8cb4d57ee2bf9