Subject: [PATCH] [WIP,Dsl] Deeper copy when populating from subprojects
When we populate a subproject's initial properties from the contents of
a subprojects (or allprojects) block, we must be careful to copy
anything that might be mutated (for example, when adding individual
elements from a subproject's own configuration).

Define a new interface to retrieve a block's
PropertiesElementDescription, and use it when populating to create a
copy of the block.

WIP because:
- needs to be implemented for the N other blocks, not just dependencies;
- traversal needs to be recursive and bottom out only at literals, but
  that's a bit complicated (e.g. multiple dependencies in a single
  configuration);
- tests.

Bug: 147139838
Test: needed
Change-Id: Ie8dd1f5c3d764b73da14581d618a592f0285dc71
(cherry picked from commit 50a6ddaf8c8751185091a7ff8c0f16e7eaad83d8)

IJ-MR-19400

GitOrigin-RevId: 01666470e84cd20e2ee04a26f288dcdd3e6fc4e0