Introduce AndroidJavaDebugEnvironment
AndroidJavaDebugEnvironment describes debug process
we are about to start.

DebugEnvironment is only thing that is needed to start debug process, but we used to create instances of many classes that we don't actually use (Runners/ExecutionEnvironment/RunProfiles) and pretend we start configuration( AndroidJavaDebugger.attachToClient) or that we have an ExecutionEnvironment (ConnectJavaDebuggerTask.launchDebugger) in order to make platform indirectly create DebugEnvironment for us.

It's much easier and cleaner to directly create DebugEnvironment and use it directly with
`DebuggerManager.attachVirtualMachine(debugEnvironment)` rather than defer it to the platform.

Additionally, it allows us to decouple code from run configurations as we don't need any to create AndroidJavaDebugEnvironment.

See go/decouple-android-debuggers-from-android-configurations

Bug: 214519121
Test: tested within next CL
Change-Id: I7fa7fe578ada26e24036459ada6cc987e25c4024