Change BenchmarkLinks to Open instead of Create
The previous change worked with the studio profliers
because the profilers open the tool window inside the
creation. This however is not how the API is intended
to be implemented. As such when trying to open a file
that is not handled by the profilers. The editor would
be created but never opened. This change updates the
call to explicitly open the editor instead of only
create it.

Test: manual, benchmark, and baseline
Fixes: 201405014
Change-Id: Ie7173ef9c38f8525cf32545f7f7db6a145e0867c
(cherry picked from commit 2df83860e9b249ce9c628542c1efa6c57d2b3ea8)