Use device dimensions when device frame is not active
This change makes the widthDp and heightDp parameters in the @Preview
annotation to impact the device Configuration screen size and not only
the layout size of the ComposeViewAdapter.
When showDecorations is true and the full device is visible, widthDp
and heightDp will apply to the ComposeViewAdapter width and height as
before.

Fixes: 196832760
Test: Added PreviewElementConfigurationTest.testSpecificDeviceSizes
Change-Id: I4f60552928d8e7ff8a71eaadeabe86e9af5bffa3