Fix failing test `com.android.tools.idea.res.ResourceFolderRepositoryTest.testFileWithNoPsi` in IDEA
This change does not spoil the test idea, because `DocumentChangeListener#documentChanged` is invoked during `document.setText`. While it is hard to set up environment to reliably get `myPsiDocumentManager.getCachedPsiFile(document) == null` in `AndroidFileChangeListener.MyDocumentListener#documentChanged` in IDEA, the second part of the test: `assertTrue(repository.hasResources(RES_AUTO, ResourceType.STRING, "from_templates"));` should succeed.

The difference is that in AS PsiFile is initialized in the DocumentChangeListener, and it sees new text. In IDEA PsiFile captures text from old document because it is initialized by LSP extensions much earlier. Later DocumentCommitThread will commit the document and AndroidFileChangeListener.PsiChangeListener will be notified, and contents of the repository will be updated. To avoid race conditions in test we invoke commit explicitly.

GitOrigin-RevId: db4c4e9b8d23540a8ee32643cc99d6440f86ba40