Fix lint which was broken as a result of MPSS
Module per source set caused lint to get duplicate paths while building
a map of directory to LintProject, this failed an assertion and caused
lint to crash.

Instead of obtaining the dir for java modules from the linked project
path, get them from the first content root. This guarantees that they
will be unique.

Modules do not really have an associated directory anymore and are a
collection of content roots. A long term fix would be to address this
and stop using a map from directories to LintProjects. This change does
not attempt to address the underlying issue but instead just ensures the
lint is usable and doesn't crash on java modules.

Bug: 216291399
Test: test added in I588cf0ceda5
Change-Id: I3d51cbd9d8676386d45ccfb71bd37a322b353ce9

GitOrigin-RevId: eb1f52fa94d369e61fd5799cefed48e5b980fc73