Fix canZoomToFit to zoom(ZoomType.FIT) behavior
ZoomToFitAction calls zoom(ZoomType.FIT) when activated, and checks for
canZoomToFit() to set its enabled state. Therefore, both functions need
to be aligned otherwise the action might be disabled when the user could
zoom-to-fit, or enabled when they could not.

Since zoom(ZoomType.FIT) sets the scale to getFitScale(false), that's
what we should use as a baseline in canZoomToFit().

Fixes: 215759823
Test: Visual by zooming to fit.
Change-Id: I005deb3d99b9c59c771dda79123c941146d40fe5