[kotlin] Decouple Kotlin bundled compiler versioning from the analyzer's compiler
Introduce 'ideCompilerVersion' and 'standaloneCompilerVersion' constants, use them accordingly instead of a wide range of constants coming from the bundled Kotlin compiler.

As we are going to decouple versioning for the Kotlin compiler and the IDE plugin, we can't use the compiler's 'KotlinVersion' anymore. So now there are two separate constants.

GitOrigin-RevId: 0826ba345cfa6b8b72e0770b70566671659b12cc