LE: Implement offset tracking.
As part of partial recompose, we need to remember the original start
and end offset of a given function. We use that information to
invalidate the compose state of that composable in order to recompose.

This CL implement what is needed to track those offset values and
not yet invoke the API with it.

Bug: 208341281
Test: Added in BasicCompileTest
Change-Id: Ib2fb16e7a6fb8dfdff25516f7380dd1eeea48822