Hide tooltips when device dropdown menu is invisible
The device dropdown menu becomes invisible when an item is selected,
or the menu is canceled by clicking outside the menu. The tooltips of
menu item may still appear on screen because of the delay of showing.

This change force hide all potentail tooltips when dropdown menu is no
longer visible on the screen.

Fixes: 207534939
Test: manually (UI-related)
Change-Id: I050cbf48d4eae8ab40cbfaa3ddaf204f2b071a6d