Refactor Compose Preview toolbar actions enable state
Add the `isAvailable` property to the actions constructor. When this
property returns false, the action is unavailable and must be disabled.
The availability function is now defined in PreviewSurfaceActionManager
and passed down when to each action when it's created. This way we can
guarantee that all the actions are consistent in terms of availability.

This refactoring makes it easier to disable these actions in other
situations, such as when there are errors to render the SceneView,
which will be part of an upcoming UX change.

Bug: 203426144
Test: N/A. Minor refactoring. Existing tests should pass.
Change-Id: I23407e558acb41c657d6093d9260410b40ee9f6a