Split Dolphin perf tests in two shards
Currently, either V1 or V2 observes high 'Initial' metrics time, while
the other has low values for it. This seems to occurs because those
metrics are recorded from the loading and first sync of the project
(which doesn't seem to be smoothed by loading/syncing the sample
project a few times before), so the first test that makes load/sync it
get worst 'Initial' metrics recorded.

This is bad, because one it makes the dashboard graphs flips based on
which (V1 or V2) are running first on some non-deterministic order, but
even if we could fix V1 to run before V2 for example, it would look like
'Initial' for V1 is worst than for V2 but that is not really true is
just an artifact of interference between both tests.

Bug: 215189237
Test: n/a
Change-Id: Ibba3384ddf85c3820088348b8cbe5d305a84d2d6