Add an utility method to check if there are render errors
Add a helper method to LayoutlibSceneManagerUtils so we can check if a
layour or Compose Preview has failed to render. This will be useful to
create error panels in the future.

This is a resubmission of change Ibffd2d, addressing the flakiness by
waiting for the refresh finish before checking the SceneViewPeerPanels.

Bug: 203426144
Test: RenderErrorTest [100 runs (locally), no failures]
Change-Id: I3fc27a44187ae210d625197b1c58354b6359e168