Paint recomposition count with the selected node
The recomposition count will be shown to the right of the selected
node next to the name of the node. If there is no room for both the
name and the count: omit the count

Also: make the image in testPaint bigger such that we can see both
the label and the recomposition count being painted.

Also: remove the use of CheckUtil.assertImageSimilar for
DeviceViewContentPanelTest which uses a different golden image per
platform. Instead:
- use the font from ImageDiffTestUtil.getDefaultFont()
- add font aliasing to draw method in DeviceViewContentPanel

After this change the images compare almost equal with running with
bazel. When running with IJ on Linux I still see differences that I
cannot explain.

Bug: 172894679
Test: Added unit test
Change-Id: Ib8e3a34468197b39ac9ce9bc3283d7afe5bbeeb4