Update device preview according to device setting.
There are two device previews: one in the AVD's
choose-a-device-definition screen, and the other one is in the
clone/edit dialog. The previews are in theory supposed to reflect the
edits users make to the devices. However, the former preview does not
update with changes in device's orientation.

The fix involves removing a field whose sole purpose is to prevent the
preview from updating to the correct orientation when the width and
height are "pretty similar".

Test: manual.
Fixes: 215982384
Change-Id: Ia93972b9383f8dfac893d3c722f3ba2722f0f471