Clarify editor banner in Version Catalog projects
When showing the banner in an editor context (which only happens for
editors on Gradle build files and .versions.toml files), clarify that
tools that might not work according to expectations will principally
be the editor tools, rather than arbitrary IDE tools.  (The boundary
is a bit blurry, as some refactorings or assistants may as part of
their operation attempt to e.g. add or remove dependencies from build
files, which might go awry, but we only have a limited character
budget for this notification and this text is probably better.

Fixes: 224833701
Test: manual (string change only)
Change-Id: I13de4c97c19ef70dbc4034db88f6319a83674f74
(cherry picked from commit c8265e338c342c7e8edf888eb8240e269f180f9f)