Cleanup Sync Model fetching
Move the decision making of whether to run an action in parallel or not
to the action level itsellf.

This simplifies the code and allows us to have one Action runner that
would run action in parallel when they can be run, and sequentially when
they cannot be run in parallel.

Bug: 179156667
Test: existing
Change-Id: Iad092a4f55407b7fc6e3adb1d7087d3561d6248a