Abstract IDE dependency models in AndroidGradleTestUtils
In the long term we do not want tests relying on light sync to depend on
the implementation details such as IDE models. These tests are generally
supposed to be project system agnostic and use the Gradle project system
as a means to set up a production-like test project.

This change begins with dependency models as they they are being
reworked and otherwise these tests will need to be changed to reflect
the new approach.

Bug: n/a
Test: n/a
Change-Id: Ib5a2e615e7bb3a99d12cf3a6895345720e3ffb70