[BA] check for min gradle version in Downloads Analyzer
File download events were added to TAPI only in 7.3.
We need to distinguish the cases when no downloads happen during
the build and when gradle simply does not provide this data.
To ensure the most actual data is used we first check if AGP version
reported from build is at least 7.2 as it requires Gradle 7.3.3.
If this is not true for any reason, check the Studio's impression
of project gradle version using
GradleVersions.getGradleVersion(Project).

Bug: 218505954
Test: added cases to DownloadsAnalyzerTest
Change-Id: Ia0d2136aa254d497baf822323b2352a4de5acbb8