[AUA] Remove some methods from AssistantInvoker
The project service is used for two reasons: one is to provide an
interface for Upgrade Assistant client code to use without introducing
a compile-time dependency; the other is to have a way to
introspect (in a limited way) the operation of the Upgrade Assistant
in tests.

In order to break dependencies, remove the methods with
AgpUpgradeRefactoringProcessor as part of the interface from this
service.  This allows the AssistantInvoker to be used purely to break
the compile-time dependency between the IDE/Lint integration and the
upgrade module.  Methods involving accepting or returning a
refactoring processor are moved to a new service internal to the
upgrade module.

Bug: n/a
Test: existing tests adjusted
Change-Id: I6181329d51963fc62332db5cdb62782117e2f37a