[kt-* only] Revert changes outside Kotlin plugin (ErrorUtils for types)
This commit is for kt-[212|213|221]-* branches. Idea since 222 has these changes.
(in master it is allowed to change idea sources, obviously)

Initial commit with changes outside the Kotlin plugin: f53c2114c38f856bdf0ff4d01de9fffeeeb630f2

There is still changes in safeargs tests, but due to the fact, that it is difficult to restore
compatibility for them and they just test, I've decided to leave them "as is"

GitOrigin-RevId: 2a0136f5bbaecc472ed5d273f694b9d22856618f