[Dsl,.toml] Take more care with plugin aliases
Handle GradleDslMethodCall forms in more places, as seen in Groovy
build files with method call syntax.  Also, code defensively around
the existence or otherwise of dependencies in the element we identify
as the alias, in case things change elsewhere once more.

Return oldElement, not newElement, from replace(): the model's element
stays the same, as we modify the contents of referred-to elements
instead.

Bug: 227793240
Test: added
Change-Id: Id32b14dce5c6816627e28f2a9f13d8e8fac9ae10