Unify ui-tooling(-preview) namespaces in Compose Preview
Before Compose 1.0.0 was released, we've made a decision to split the
ui-tooling module into two: ui-tooling and ui-tooling-preview. Since
there were quite a few compose projects at that point using alpha and
beta releases of 1.0.0, we created a mechanism in Android Studio to
support both scenarios (single or split library).

This has now become problematic, because one of the assumptions we've
made when calculating which scenario was being used was to check for
the presence of the ui-tooling-preview **artifact**. However, now that
we're aiming to support Compose Previews in androidx, ui-tooling-preview
can now be added as a project dependency, breaking this assumption and
making us unable to load ComposeViewAdapter.

Instead of fixing this behavior, we can go further and remove the single
ui-tooling library support. Considering that Compose 1.1.1 is stable, we
shouldn't need to support pre-stable Compose 1.0.0 at this point.

Bug: 225192013
Test: compose-designer tests
Change-Id: I34350590d7306b4c3dacdff387a00cc1a043282b