Fix user gradle.properties file lookup in GradleUtil
GradleProperties.getUserGradlePropertiesFile estimates location of
user gradle.properties file based on system properties and environment
variables, but it does not take into account IDE project settings.
Change this in GradleUtil, check gradle execution settings for this
project first. This will make sure that resolved path is the same as
will be used for running gradle.
E.g. if Gradle home was changed in Gradle settings page it will not
be propagated to system properties and environment variable,
so we can get it from Gradle execution settings for the project only.

If we can not get Gradle home path from the execution settings
fallback to usual assumptions from system properties and variables.

Bug: 228462473
Test: Added GradleUtilAndroidGradleTest
  .testUserGradlePropertiesFileDetectionForGradleHomeChangedInSettings
Change-Id: Ic2c2aa26ac816d3545fb1a259d594eb4e2314d4d