Remove LARGE_MACHINE from targets
The targets using larger VMs are no longer the majority of critical path
targets. To simplify the pool setup, we are going to move all tests back
to the default pool.

The shard count for two split targets is increased since these targets
start to time out. Having an extra shard for this target is
not a problem, since using a single pool should bring better sharing of
machines.

This reverts commit c6b3f6c1abeb74d61fc71325a30c3744a7a9ea2c.

Reason for revert: Rolling forward and adding more shards to targets
that have test timeouts. It appears timeout failures were mostly
on coverage targets
https://screenshot.googleplex.com/9FePfbkRvrGXAHm

Bug: 165823108
Test: n/a
Change-Id: I065af319d308b059f6880a94c62cde3818b94571