[Sync] Adjust a test
The error message emitted when running an incompatible preview version
of AGP in a preview version of Studio differs from the message emitted
when that preview version of AGP is run by a released version of
Studio.  Alter the test for that error message to itself behave
differently depending on whether the test is run in the context of a
preview or release version.

Bug: none filed
Test: fixed
Change-Id: I2c957766b6d41c57e7f3bcba94fc8a317ae22a18