Fix attaching native models when deploying
When deploying an app with native components, Studio
would like to have the debug information for the ABI of the deployment
devices

With V2 native sync (since AGP 4.2) these are always avaliable
and are directly read from files. With V1 native sync these
might not be avaliable, and Studio currently triggers a sync
during deployment in those cases.

This change fixes attaching those models correctly so that both
 1. The functionality the model fetch was meant to enable can work
 2. The model fetch is not repeatedly triggered for subsequent
    deployments.

Bug: 226372843
Test: Fixed and strengthened assertions in MakeBeforeRunTaskProviderIntegrationNativeSyncV1Test
Change-Id: I5510a232c7d15ef3506e0875bbe2886d8eede8e5