Android sync requires an IDE project
but should probably not crash if invoked this way since third party code
may attempt to resolve a project in a different context.

In Android Studio, `ExternalSystemTaskId` is instantiated form a live
`Project` only and is supposed to be always resolvable. Therefore, we
report an error but do not crash if it happens.

Bug: 228441874
Test: n/a
Change-Id: If82b32de1206dc699cd2519f3c748c0d80b63514