[AUA] Rename a class
The AgpClasspathDependencyRefactoringProcessor has for a while had to
do substantially more than looking for classpath dependencies: normal
dependencies in buildSrc, and plugins Dsl declarations in both Gradle
build files and settings files also need to be inspected.  Make the
name of the processor more explicitly connected to its function, and
adjust UI strings slightly for the same reason.  Make test, variable
and data directory names match the new class name.

For backward-compatibility reasons we have to preserve the
studio_stats name for the processor component.

Bug: 231690925
Test: existing (no change in functionality)
Change-Id: Ifb3460782129d8795da6e8ecc2adb50d33d78ab0